+++ aws/init.go
@@ -29,7 +29,7 @@ import (
 )
 
 var (
-	AccessService, InfraService, StorageService, NotificationService, QueueService, DnsService cloud.Service
+	AccessService, InfraService, StorageService, NotificationService, QueueService, DnsService, LambdaService cloud.Service
 )
 
 func InitSession(region, profile string) (*session.Session, error) {
@@ -69,6 +69,7 @@ func InitServices(conf map[string]interface{}, log *logger.Logger) error {
 	NotificationService = NewNotification(sess, awsconf, log)
 	QueueService = NewQueue(sess, awsconf, log)
 	DnsService = NewDns(sess, awsconf, log)
+	LambdaService = NewLambda(sess, awsconf, log)
 
 	cloud.ServiceRegistry[InfraService.Name()] = InfraService
 	cloud.ServiceRegistry[AccessService.Name()] = AccessService
@@ -76,6 +77,7 @@ func InitServices(conf map[string]interface{}, log *logger.Logger) error {
 	cloud.ServiceRegistry[NotificationService.Name()] = NotificationService
 	cloud.ServiceRegistry[QueueService.Name()] = QueueService
 	cloud.ServiceRegistry[DnsService.Name()] = DnsService
+	cloud.ServiceRegistry[LambdaService.Name()] = LambdaService
 
 	return nil
 }